src/models/element_detector.py
"""Element detection models wrapper"""
import numpy as np
from typing import List, Dict, Optional, Tuple


class ElementDetector:
    """Wrapper for different element detection models"""

    def __init__(self, model_type: str = 'pattern'):
        self.model_type = model_type
        self.model = None

    def detect_elements(self, image: np.ndarray, element_types: List[str] = None) -> List[Dict]:
        """Detect UI elements in image"""
        if self.model_type == 'pattern':
            # Use traditional computer vision
            return self._pattern_detection(image)
        elif self.model_type == 'yolo-world':
            return self._yolo_detection(image, element_types)
        else:
            return []

    def _pattern_detection(self, image: np.ndarray) -> List[Dict]:
        """Pattern-based detection using OpenCV"""
        import cv2

        elements = []
        gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)

        # Detect buttons (rectangles with text)
        edges = cv2.Canny(gray, 50, 150)
        contours, _ = cv2.findContours(edges, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)

        for contour in contours:
            x, y, w, h = cv2.boundingRect(contour)
            if w > 50 and h > 20 and w < 500 and h < 100:  # Button-like dimensions
                elements.append({
                    'type': 'button',
                    'bounds': {'x': x, 'y': y, 'width': w, 'height': h},
                    'confidence': 0.7
                })

        return elements

    def _yolo_detection(self, image: np.ndarray, element_types: List[str]) -> List[Dict]:
        """YOLO-based detection"""
        # Implementation depends on YOLO model
        # This is a placeholder
        return []