requirements.txt
playwright==1.40.0
pyyaml==6.0.1
click==8.1.7
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2
Pillow==10.1.0
opencv-python-headless==4.8.1.78
ultralytics==8.0.200  # For YOLO-World
numpy==1.24.4
requests==2.31.0
jinja2==3.1.2
colorama==0.4.6
python-dotenv==1.0.0
spacy==3.7.2  # Optional for advanced NLP
aiohttp==3.9.1
jsonpath-ng==1.6.0
jinja2==3.1.2
jsonschema==4.20.0